+++ tests/test_registry.py
@@ -1,11 +1,11 @@
 import pytest
 from pydantic import BaseModel
 
-import graphene_pydantic.objecttype.registry as registry
-from graphene_pydantic import PydanticObjectType
-from graphene_pydantic.objecttype.registry import (
+import graphene_pydantic.registry as registry
+from graphene_pydantic import PydanticInputObjectType, PydanticObjectType
+from graphene_pydantic.registry import (
     Registry,
-    assert_is_pydantic_object_type,
+    assert_is_correct_type,
     get_global_registry,
     reset_global_registry,
 )
@@ -22,37 +22,39 @@ def _get_dummy_classes():
     return (Foo, GraphFoo)
 
 
-def test_assert_is_pydantic_object_type():
+def test_assert_is_correct_type():
     Foo, GraphFoo = _get_dummy_classes()
     with pytest.raises(TypeError) as exc:
-        assert_is_pydantic_object_type(Foo)
+        assert_is_correct_type(Foo, PydanticObjectType)
 
-    assert "Expected PydanticObjectType, but got" in exc.value.args[0]
-
-    assert_is_pydantic_object_type(GraphFoo)
+    assert_is_correct_type(GraphFoo, PydanticObjectType)
 
 
 def test_get_global_registry():
-    assert isinstance(get_global_registry(), Registry)
+    assert isinstance(get_global_registry(PydanticObjectType), Registry)
+
+
+def test_get_global_registry_for_input():
+    assert isinstance(get_global_registry(PydanticInputObjectType), Registry)
 
 
 def test_reset_global_registry():
-    get_global_registry()
-    assert registry.registry is not None
-    reset_global_registry()
-    assert registry.registry is None
+    get_global_registry(PydanticInputObjectType)
+    assert registry.registry.get(PydanticInputObjectType) is not None
+    reset_global_registry(PydanticInputObjectType)
+    assert registry.registry.get(PydanticInputObjectType) is None
 
 
 def test_register_and_get_type_for_model():
     Foo, GraphFoo = _get_dummy_classes()
-    r = get_global_registry()
+    r = get_global_registry(PydanticObjectType)
     r.register(GraphFoo)
     assert r.get_type_for_model(Foo) is GraphFoo
 
 
 def test_register_object_field_and_get_for_graphene_field():
     Foo, GraphFoo = _get_dummy_classes()
-    r = get_global_registry()
+    r = get_global_registry(PydanticObjectType)
     field = r.get_object_field_for_graphene_field(GraphFoo, "name")
     assert field is not None
     assert field.type_ == str
